Fix some hairiness in tabasco with complex arrays.
Also get some fasls out of the repo.  Oops.

pkhuong pointed out to me (on #lisp) that the "portable" tabasco sort as
written is incorrect, due to the lack of sbcl's with-array-data.

lisp arrays are tricky things beyond the simple case (where it's just
arefs).  But I think I've got it right now: displaced arrays and
adjustable arrays can be treated normally, but arrays with fill-pointers
have to be distinguished (apparently at runtime, since even simple-array
doesn't necessarily exclude arrays with fill pointers) since SORT should
only work on those elements of theres that are visible without the
fill-pointer.

SBCL "just" has a with-array-data macro that relies on its internal
representation of complex arrays as having an underlying simple array,
but this obviously isn't portable.

I really need to test these things for speed now.  And correctness, of
course, but things like this are pretty subtle.

Oh, yes, pkhuong also pointed out that in the length one case the other
arguments still need to be evaluated, so I fixed that.